Reworked the structure.
