+++ src/ModelItem/Extensions/PropertyCategoryExtensions.cs
@@ -1,5 +1,4 @@
-﻿using PedramElmi.Navisworks.Toolkit.ModelItem;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using Autodesk.Navisworks.Api;
 
@@ -7,10 +6,8 @@ namespace PedramElmi.Navisworks.Toolkit
 {
     public static class PropertyCategoryExtensions
     {
-        #region Public Methods
-
         /// <summary>
-        /// Returns the categories display name
+        /// Returns the categories DisplayName
         /// </summary>
         /// <param name="categories">
         /// </param>
@@ -18,7 +15,8 @@ namespace PedramElmi.Navisworks.Toolkit
         /// </returns>
         public static IEnumerable<string> GetCategoriesDisplaName(this IEnumerable<PropertyCategory> categories)
         {
-            return CategoriesProperties.GetCategoriesDisplaName(categories);
+            return (from category in categories
+                    select category.DisplayName).Distinct();
         }
         
         /// <summary>
@@ -30,9 +28,17 @@ namespace PedramElmi.Navisworks.Toolkit
         /// </returns>
         public static IEnumerable<string> GetPropertiesDisplayName(this PropertyCategory category)
         {
-            return CategoriesProperties.GetPropertiesDisplayName(category);
+            return from property in category.Properties select property.DisplayName;
         }
 
-        #endregion Public Methods
+        public static IDictionary<string, IDictionary<string, object>> ToDictionary(this IEnumerable<PropertyCategory> categories)
+        {
+            var dictionary = new Dictionary<string, IDictionary<string, object>>();
+            foreach (var category in categories)
+            {
+                dictionary.Add(category.DisplayName, category.Properties.ToDictionary());
+            }
+            return dictionary;
+        }
     }
 }
